+++ source/tests/test.python/test_renaming.py
@@ -4,8 +4,7 @@
 
 import json
 
-from omni.scene.optimizer.core.scripts import standalone
-
+from .scripts import standalone
 from .test_utils import Test_Operation