+++ source/validators/premake5.lua
@@ -5,7 +5,7 @@ project_with_location("validators_python")
 
     kind "Utility"
 
-    so_build.python_module({
-        module_path = "omni/scene/optimizer/validators",
-        python_sources = "python/omni/scene/optimizer/validators/*.py",
+    so_build.symlink_folder({
+        target_dir = "python/omni/scene/optimizer/validators",
+        source_dir = "python/omni/scene/optimizer/validators",
     })